scripts/requirements.txt
gensim
jieba
python-docx
